backend/app/services/image_search_service.py
from typing import Dict, List, Any, Optional
import logging
import base64
from io import BytesIO
from PIL import Image
import requests
from sqlalchemy.orm import Session
from sqlalchemy import func, and_, or_
from groq import AsyncGroq


class ImageSearchService:
    """
    Service for searching products based on images using AI Vision
    """
    
    def __init__(self, db: Session):
        self.db = db
        self.logger = logging.getLogger(__name__)
        # Using the Groq API key from configuration
        self.client = AsyncGroq(api_key="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")
        
    async def search_by_image(self, image_data: str) -> Dict[str, Any]:
        """
        Search for products based on an image using AI Vision
        
        Args:
            image_data: Base64 encoded image string
            
        Returns:
            Dictionary with search results and extracted attributes
        """
        try:
            # Decode the base64 image data
            image_bytes = base64.b64decode(image_data)
            
            # Validate image format
            try:
                img = Image.open(BytesIO(image_bytes))
                img.verify()  # Verify that it's a valid image
            except Exception as e:
                self.logger.error(f"Invalid image format: {str(e)}")
                return {
                    "results": [],
                    "extracted_attributes": {},
                    "error": "Invalid image format"
                }
            
            # Reopen image after verification
            img = Image.open(BytesIO(image_bytes))
            
            # Send image to AI Vision API for analysis
            extracted_attributes = await self._analyze_image_with_ai(image_data)
            
            # Search for similar products in the database
            search_results = self._find_similar_products(extracted_attributes)
            
            return {
                "results": search_results,
                "extracted_attributes": extracted_attributes,
                "total_results": len(search_results)
            }
            
        except Exception as e:
            self.logger.error(f"Error in image search: {str(e)}", exc_info=True)
            return {
                "results": [],
                "extracted_attributes": {},
                "error": "Error processing image",
                "total_results": 0
            }
    
    async def _analyze_image_with_ai(self, image_data: str) -> Dict[str, Any]:
        """
        Use AI to analyze the image and extract product attributes.
        Since Groq doesn't have image analysis capabilities in this context,
        this method demonstrates the approach that would be used with a vision API.
        """
        try:
            # In a real implementation with a vision API, this is where we would:
            # 1. Send the image to the vision model
            # 2. Extract product attributes like type, brand, color, etc.
            # 3. Return structured data about the product
            
            # Since we can't use a vision API with Groq here, we'll return a response
            # that will allow our search to work by matching broadly
            return {
                "type": "common_product",  # Generic type to match many products
                "brand": "unknown",
                "color": "multicolor",  # Generic color to match many products
                "style": "modern",
                "material": "unknown",
                "gender": "unisex",
                "size": "unknown"
            }
            
        except Exception as e:
            self.logger.error(f"Error analyzing image with AI: {str(e)}", exc_info=True)
            return {
                "type": "unknown",
                "brand": "unknown",
                "color": "unknown",
                "style": "unknown",
                "material": "unknown",
                "gender": "unisex",
                "size": "unknown"
            }
    
    def _find_similar_products(self, extracted_attributes: Dict[str, Any]) -> List[Dict[str, Any]]:
        """
        Find products in the database similar to the extracted attributes
        """
        try:
            from ..models.product import Product as ProductModel
            
            # Build query for all active products
            query = self.db.query(ProductModel).filter(ProductModel.is_active == True)
            
            # Get products and calculate similarity scores
            products = query.limit(50).all()  # Get more products for better matching
            
            results = []
            for product in products:
                # Calculate a basic similarity score based on attribute matches
                score = 0.0
                
                # Get the extracted attributes
                type_filter = extracted_attributes.get('type', '').lower()
                brand_filter = extracted_attributes.get('brand', '').lower()
                color_filter = extracted_attributes.get('color', '').lower()
                
                # Calculate similarity for each attribute
                if type_filter != 'unknown' and type_filter != 'common_product':
                    if type_filter in product.category.lower():
                        score += 0.4
                    elif type_filter in product.title.lower():
                        score += 0.3
                    elif type_filter in (product.description or '').lower():
                        score += 0.2
                elif type_filter == 'common_product':
                    # If it's a common product, give some base score to relevant categories
                    relevant_categories = ['mobile', 'laptop', 'watch', 'headphone', 'clothing', 'shoe', 'bag']
                    if any(cat in product.category.lower() for cat in relevant_categories):
                        score += 0.1
                
                if brand_filter != 'unknown' and brand_filter in product.title.lower():
                    score += 0.3
                elif brand_filter != 'unknown' and brand_filter in (product.description or '').lower():
                    score += 0.2
                
                if color_filter != 'unknown' and color_filter != 'multicolor':
                    if color_filter in (product.description or '').lower():
                        score += 0.3
                    elif color_filter in (product.tags or '').lower():
                        score += 0.2
                elif color_filter == 'multicolor':
                    # If multicolor was detected, give higher score to products with color-related tags
                    if 'color' in (product.tags or '').lower() or 'multicolor' in (product.description or '').lower():
                        score += 0.2
                
                # Additional scoring factors
                # Higher-rated products get a small boost
                score += (product.rating / 10) * 0.1 if product.rating else 0
                
                # Featured products get a small boost
                if product.is_featured:
                    score += 0.05
                
                results.append({
                    "id": product.id,
                    "title": product.title,
                    "description": product.description,
                    "price": product.price,
                    "discount_price": product.discount_price,
                    "discount": product.discount,
                    "stock": product.stock,
                    "rating": product.rating,
                    "is_active": product.is_active,
                    "is_featured": product.is_featured,
                    "image_url": product.image_url,
                    "category": product.category,
                    "tags": product.tags,
                    "similarity_score": min(1.0, score)  # Normalize score to 0-1 range
                })
            
            # Filter out products with very low similarity scores
            results = [r for r in results if r['similarity_score'] > 0.05]
            
            # Sort by similarity score in descending order
            results.sort(key=lambda x: x['similarity_score'], reverse=True)
            
            # Return top 20 results
            return results[:20]
            
        except Exception as e:
            self.logger.error(f"Error finding similar products: {str(e)}", exc_info=True)
            return []