backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
jinja2==3.1.2
requests==2.31.0
alembic==1.13.1
pytest==7.4.3
httpx==0.25.2
pillow
python-slugify==8.0.1
phonenumbers==8.13.28
pyjwt==2.8.0
cachetools==5.3.2
tiktoken
openai==1.3.4
groq>=0.4.1
google-generativeai==0.4.0